+++ contracts/data-storage/andromeda-counter/src/contract.rs
@@ -12,7 +12,7 @@ use andromeda_data_storage::counter::{
 use andromeda_std::{
     ado_base::{InstantiateMsg as BaseInstantiateMsg, MigrateMsg},
     ado_contract::ADOContract,
-    common::{actions::call_action, context::ExecuteContext, encode_binary},
+    common::{context::ExecuteContext, encode_binary},
     error::ContractError,
 };
 use cw_utils::nonpayable;
@@ -88,16 +88,16 @@ pub fn execute(
     }
 }
 
-fn handle_execute(mut ctx: ExecuteContext, msg: ExecuteMsg) -> Result<Response, ContractError> {
+fn handle_execute(ctx: ExecuteContext, msg: ExecuteMsg) -> Result<Response, ContractError> {
     let action = msg.as_ref().to_string();
 
-    let action_response = call_action(
-        &mut ctx.deps,
-        &ctx.info,
-        &ctx.env,
-        &ctx.amp_ctx,
-        msg.as_ref(),
-    )?;
+    // let action_response = call_action(
+    //     &mut ctx.deps,
+    //     &ctx.info,
+    //     &ctx.env,
+    //     &ctx.amp_ctx,
+    //     msg.as_ref(),
+    // )?;
 
     let res = match msg {
         ExecuteMsg::Increment {} => execute_increment(ctx, action),
@@ -115,10 +115,11 @@ fn handle_execute(mut ctx: ExecuteContext, msg: ExecuteMsg) -> Result<Response, 
         _ => ADOContract::default().execute(ctx, msg),
     }?;
 
-    Ok(res
-        .add_submessages(action_response.messages)
-        .add_attributes(action_response.attributes)
-        .add_events(action_response.events))
+    Ok(
+        res, // .add_submessages(action_response.messages)
+            // .add_attributes(action_response.attributes)
+            // .add_events(action_response.events))
+    )
 }
 
 pub fn execute_increment(ctx: ExecuteContext, action: String) -> Result<Response, ContractError> {