+++ contracts/data-storage/andromeda-counter/src/mock.rs
@@ -1,11 +1,13 @@
 #![cfg(all(not(target_arch = "wasm32"), feature = "testing"))]
 use crate::contract::{execute, instantiate, query};
-use andromeda_data_storage::counter::{CounterRestriction, ExecuteMsg, InstantiateMsg, QueryMsg};
+use andromeda_data_storage::counter::{
+    CounterRestriction, ExecuteMsg, InstantiateMsg, QueryMsg, State,
+};
 use andromeda_data_storage::counter::{
     GetCurrentAmountResponse, GetDecreaseAmountResponse, GetIncreaseAmountResponse,
     GetInitialAmountResponse, GetRestrictionResponse,
 };
-use andromeda_std::ado_base::rates::{Rate, RatesMessage};
+// use andromeda_std::ado_base::rates::{Rate, RatesMessage};
 use andromeda_testing::mock::MockApp;
 use andromeda_testing::{
     mock_ado,
@@ -18,6 +20,7 @@ pub struct MockCounter(Addr);
 mock_ado!(MockCounter, ExecuteMsg, QueryMsg);
 
 impl MockCounter {
+    #[allow(clippy::too_many_arguments)]
     pub fn instantiate(
         code_id: u64,
         sender: Addr,
@@ -185,9 +188,11 @@ pub fn mock_counter_instantiate_msg(
         kernel_address,
         owner,
         restriction,
-        initial_amount,
-        increase_amount,
-        decrease_amount,
+        initial_state: State {
+            initial_amount,
+            increase_amount,
+            decrease_amount,
+        },
     }
 }