+++ contracts/finance/andromeda-vesting/src/testing/tests.rs
@@ -1,5 +1,7 @@
 use andromeda_std::{
-    amp::Recipient, common::withdraw::WithdrawalType, error::ContractError,
+    amp::Recipient,
+    common::{withdraw::WithdrawalType, Milliseconds},
+    error::ContractError,
     testing::mock_querier::MOCK_KERNEL_CONTRACT,
 };
 use cosmwasm_std::{
@@ -7,7 +9,6 @@ use cosmwasm_std::{
     testing::{mock_env, mock_info, MOCK_CONTRACT_ADDR},
     BankMsg, Decimal, DepsMut, Response, Uint128,
 };
-use cw_utils::Duration;
 
 use crate::{
     contract::{execute, instantiate, query},
@@ -17,14 +18,10 @@ use crate::{
 
 use andromeda_finance::vesting::{BatchResponse, Config, ExecuteMsg, InstantiateMsg, QueryMsg};
 
-const UNBONDING_BLOCK_DURATION: u64 = 5;
-
 fn init(deps: DepsMut) -> Response {
     let msg = InstantiateMsg {
         recipient: Recipient::from_string("recipient"),
-        is_multi_batch_enabled: true,
         denom: "uusd".to_string(),
-        unbonding_duration: Duration::Height(UNBONDING_BLOCK_DURATION),
         kernel_address: MOCK_KERNEL_CONTRACT.to_string(),
         owner: None,
     };
@@ -35,8 +32,8 @@ fn init(deps: DepsMut) -> Response {
 
 fn create_batch(
     deps: DepsMut,
-    lockup_duration: Option<u64>,
-    release_unit: u64,
+    lockup_duration: Option<Milliseconds>,
+    release_unit: Milliseconds,
     release_amount: WithdrawalType,
 ) -> Response {
     // Create batch with half of the release_unit.
@@ -68,7 +65,6 @@ fn test_instantiate() {
     assert_eq!(
         Config {
             recipient: Recipient::from_string("recipient"),
-            is_multi_batch_enabled: true,
             denom: "uusd".to_string(),
         },
         CONFIG.load(deps.as_ref().storage).unwrap()
@@ -84,7 +80,7 @@ fn test_create_batch_unauthorized() {
 
     let msg = ExecuteMsg::CreateBatch {
         lockup_duration: None,
-        release_unit: 1,
+        release_unit: Milliseconds::from_seconds(1),
         release_amount: WithdrawalType::Amount(Uint128::zero()),
     };
 
@@ -103,7 +99,7 @@ fn test_create_batch_no_funds() {
 
     let msg = ExecuteMsg::CreateBatch {
         lockup_duration: None,
-        release_unit: 1,
+        release_unit: Milliseconds::from_seconds(1),
         release_amount: WithdrawalType::Amount(Uint128::zero()),
     };
 
@@ -126,7 +122,7 @@ fn test_create_batch_invalid_denom() {
 
     let msg = ExecuteMsg::CreateBatch {
         lockup_duration: None,
-        release_unit: 1,
+        release_unit: Milliseconds::from_seconds(1),
         release_amount: WithdrawalType::Amount(Uint128::zero()),
     };
 
@@ -149,7 +145,7 @@ fn test_create_batch_valid_denom_zero_amount() {
 
     let msg = ExecuteMsg::CreateBatch {
         lockup_duration: None,
-        release_unit: 1,
+        release_unit: Milliseconds::from_seconds(1),
         release_amount: WithdrawalType::Amount(Uint128::zero()),
     };
 
@@ -167,7 +163,7 @@ fn test_create_batch_release_unit_zero() {
 
     let msg = ExecuteMsg::CreateBatch {
         lockup_duration: None,
-        release_unit: 0,
+        release_unit: Milliseconds::zero(),
         release_amount: WithdrawalType::Amount(Uint128::zero()),
     };
 
@@ -185,7 +181,7 @@ fn test_create_batch_release_amount_zero() {
 
     let msg = ExecuteMsg::CreateBatch {
         lockup_duration: None,
-        release_unit: 10,
+        release_unit: Milliseconds::from_seconds(10),
         release_amount: WithdrawalType::Amount(Uint128::zero()),
     };
 
@@ -203,19 +199,19 @@ fn test_create_batch() {
 
     let msg = ExecuteMsg::CreateBatch {
         lockup_duration: None,
-        release_unit: 10,
+        release_unit: Milliseconds::from_seconds(10),
         release_amount: WithdrawalType::Amount(Uint128::new(10)),
     };
 
     let res = execute(deps.as_mut(), mock_env(), info.clone(), msg).unwrap();
-    let current_time = mock_env().block.time.seconds();
+    let current_time = Milliseconds::from_seconds(mock_env().block.time.seconds());
 
     assert_eq!(
         Response::new()
             .add_attribute("action", "create_batch")
             .add_attribute("amount", "100")
             .add_attribute("lockup_end", current_time.to_string())
-            .add_attribute("release_unit", "10")
+            .add_attribute("release_unit", Milliseconds::from_seconds(10).to_string())
             .add_attribute("release_amount", "Amount(Uint128(10))"),
         res
     );
@@ -227,7 +223,7 @@ fn test_create_batch() {
             amount: Uint128::new(100),
             amount_claimed: Uint128::zero(),
             lockup_end: current_time,
-            release_unit: 10,
+            release_unit: Milliseconds::from_seconds(10),
             release_amount: WithdrawalType::Amount(Uint128::new(10)),
             last_claimed_release_time: current_time,
         },
@@ -238,8 +234,8 @@ fn test_create_batch() {
 
     // Try to create another batch.
     let msg = ExecuteMsg::CreateBatch {
-        lockup_duration: Some(100),
-        release_unit: 10,
+        lockup_duration: Some(Milliseconds::from_seconds(100)),
+        release_unit: Milliseconds::from_seconds(10),
         release_amount: WithdrawalType::Amount(Uint128::new(10)),
     };
 
@@ -249,8 +245,8 @@ fn test_create_batch() {
         Response::new()
             .add_attribute("action", "create_batch")
             .add_attribute("amount", "100")
-            .add_attribute("lockup_end", (current_time + 100).to_string())
-            .add_attribute("release_unit", "10")
+            .add_attribute("lockup_end", (current_time.plus_seconds(100)).to_string())
+            .add_attribute("release_unit", Milliseconds::from_seconds(10).to_string())
             .add_attribute("release_amount", "Amount(Uint128(10))"),
         res
     );
@@ -261,73 +257,15 @@ fn test_create_batch() {
         Batch {
             amount: Uint128::new(100),
             amount_claimed: Uint128::zero(),
-            lockup_end: current_time + 100,
-            release_unit: 10,
+            lockup_end: current_time.plus_seconds(100),
+            release_unit: Milliseconds::from_seconds(10),
             release_amount: WithdrawalType::Amount(Uint128::new(10)),
-            last_claimed_release_time: current_time + 100,
+            last_claimed_release_time: current_time.plus_seconds(100),
         },
         batch
     );
 
     assert_eq!(3, NEXT_ID.load(deps.as_ref().storage).unwrap());
-}
-
-#[test]
-fn test_create_batch_multi_batch_not_supported() {
-    let mut deps = mock_dependencies_custom(&[]);
-    let msg = InstantiateMsg {
-        recipient: Recipient::from_string("recipient"),
-        is_multi_batch_enabled: false,
-        denom: "uusd".to_string(),
-        unbonding_duration: Duration::Height(0u64),
-        kernel_address: MOCK_KERNEL_CONTRACT.to_string(),
-        owner: None,
-    };
-
-    let info = mock_info("owner", &[]);
-    instantiate(deps.as_mut(), mock_env(), info, msg).unwrap();
-
-    let info = mock_info("owner", &coins(100, "uusd"));
-
-    let msg = ExecuteMsg::CreateBatch {
-        lockup_duration: Some(100),
-        release_unit: 10,
-        release_amount: WithdrawalType::Amount(Uint128::new(10)),
-    };
-
-    let res = execute(deps.as_mut(), mock_env(), info.clone(), msg.clone()).unwrap();
-    let current_time = mock_env().block.time.seconds();
-
-    assert_eq!(
-        Response::new()
-            .add_attribute("action", "create_batch")
-            .add_attribute("amount", "100")
-            .add_attribute("lockup_end", (current_time + 100).to_string())
-            .add_attribute("release_unit", "10")
-            .add_attribute("release_amount", "Amount(Uint128(10))"),
-        res
-    );
-
-    let batch = batches().load(deps.as_ref().storage, 1).unwrap();
-
-    assert_eq!(
-        Batch {
-            amount: Uint128::new(100),
-            amount_claimed: Uint128::zero(),
-            lockup_end: current_time + 100,
-            release_unit: 10,
-            release_amount: WithdrawalType::Amount(Uint128::new(10)),
-            last_claimed_release_time: current_time + 100,
-        },
-        batch
-    );
-
-    assert_eq!(2, NEXT_ID.load(deps.as_ref().storage).unwrap());
-
-    // Try to create another batch.
-    let res = execute(deps.as_mut(), mock_env(), info, msg);
-
-    assert_eq!(ContractError::MultiBatchNotSupported {}, res.unwrap_err());
 }
 
 #[test]
@@ -355,8 +293,8 @@ fn test_claim_batch_still_locked() {
 
     // Create batch.
     let msg = ExecuteMsg::CreateBatch {
-        lockup_duration: Some(100),
-        release_unit: 10,
+        lockup_duration: Some(Milliseconds::from_seconds(100)),
+        release_unit: Milliseconds::from_seconds(10),
         release_amount: WithdrawalType::Amount(Uint128::new(10)),
     };
 
@@ -382,7 +320,7 @@ fn test_claim_batch_no_funds_available() {
     // Create batch.
     let msg = ExecuteMsg::CreateBatch {
         lockup_duration: None,
-        release_unit: 10,
+        release_unit: Milliseconds::from_seconds(10),
         release_amount: WithdrawalType::Amount(Uint128::new(10)),
     };
 
@@ -406,7 +344,7 @@ fn test_claim_batch_single_claim() {
     init(deps.as_mut());
     let info = mock_info("owner", &coins(100, "uusd"));
 
-    let release_unit = 10;
+    let release_unit = Milliseconds::from_seconds(10);
 
     // Create batch.
     let msg = ExecuteMsg::CreateBatch {
@@ -424,13 +362,13 @@ fn test_claim_batch_single_claim() {
     // Skip time.
     let mut env = mock_env();
     // A single release is available.
-    env.block.time = env.block.time.plus_seconds(release_unit);
+    env.block.time = env.block.time.plus_seconds(release_unit.seconds());
 
     // Query created batch.
     let msg = QueryMsg::Batch { id: 1 };
     let res: BatchResponse = from_json(query(deps.as_ref(), env.clone(), msg).unwrap()).unwrap();
 
-    let lockup_end = mock_env().block.time.seconds();
+    let lockup_end = Milliseconds::from_seconds(mock_env().block.time.seconds());
     assert_eq!(
         BatchResponse {
             id: 1,
@@ -466,16 +404,16 @@ fn test_claim_batch_single_claim() {
             .add_attribute("amount_left", "90"),
         res
     );
-    let lockup_end = mock_env().block.time.seconds();
+    let lockup_end = Milliseconds::from_seconds(mock_env().block.time.seconds());
 
     assert_eq!(
         Batch {
             amount: Uint128::new(100),
             amount_claimed: Uint128::new(10),
             lockup_end,
-            release_unit: 10,
+            release_unit: Milliseconds::from_seconds(10),
             release_amount: WithdrawalType::Amount(Uint128::new(10)),
-            last_claimed_release_time: lockup_end + release_unit,
+            last_claimed_release_time: lockup_end.plus_milliseconds(release_unit),
         },
         batches().load(deps.as_ref().storage, 1u64).unwrap()
     );
@@ -487,7 +425,7 @@ fn test_claim_batch_not_nice_numbers_single_release() {
     init(deps.as_mut());
     let info = mock_info("owner", &coins(10, "uusd"));
 
-    let release_unit = 10;
+    let release_unit = Milliseconds::from_seconds(10);
 
     // Create batch.
     let msg = ExecuteMsg::CreateBatch {
@@ -505,7 +443,7 @@ fn test_claim_batch_not_nice_numbers_single_release() {
     // Skip time.
     let mut env = mock_env();
     // A single release is available.
-    env.block.time = env.block.time.plus_seconds(release_unit);
+    env.block.time = env.block.time.plus_seconds(release_unit.seconds());
 
     // Claim batch.
     let msg = ExecuteMsg::Claim {
@@ -527,16 +465,16 @@ fn test_claim_batch_not_nice_numbers_single_release() {
             .add_attribute("amount_left", "3"),
         res
     );
-    let lockup_end = mock_env().block.time.seconds();
+    let lockup_end = Milliseconds::from_seconds(mock_env().block.time.seconds());
 
     assert_eq!(
         Batch {
             amount: Uint128::new(10),
             amount_claimed: Uint128::new(7),
             lockup_end,
-            release_unit: 10,
+            release_unit: Milliseconds::from_seconds(10),
             release_amount: WithdrawalType::Amount(Uint128::new(10)),
-            last_claimed_release_time: lockup_end + release_unit,
+            last_claimed_release_time: lockup_end.plus_milliseconds(release_unit),
         },
         batches().load(deps.as_ref().storage, 1u64).unwrap()
     );
@@ -549,7 +487,7 @@ fn test_claim_batch_not_nice_numbers_multiple_releases() {
     let vesting_amount = 1_000_000_000_000_000_000u128;
     let info = mock_info("owner", &coins(vesting_amount, "uusd"));
 
-    let release_unit = 1; // 1 second
+    let release_unit = Milliseconds::from_seconds(1); // 1 second
     let duration: u64 = 60 * 60 * 24 * 365 * 5; // 5 years
     let percent_release = Decimal::from_ratio(Uint128::one(), Uint128::from(duration));
 
@@ -569,7 +507,7 @@ fn test_claim_batch_not_nice_numbers_multiple_releases() {
     // Skip time.
     let mut env = mock_env();
     // Two releases are out.
-    env.block.time = env.block.time.plus_seconds(2 * release_unit);
+    env.block.time = env.block.time.plus_seconds(2 * release_unit.seconds());
 
     // Claim batch.
     let msg = ExecuteMsg::Claim {
@@ -591,16 +529,16 @@ fn test_claim_batch_not_nice_numbers_multiple_releases() {
             .add_attribute("amount_left", (vesting_amount - 12683916792).to_string()),
         res
     );
-    let lockup_end = mock_env().block.time.seconds();
+    let lockup_end = Milliseconds::from_seconds(mock_env().block.time.seconds());
 
     assert_eq!(
         Batch {
             amount: Uint128::new(vesting_amount),
             amount_claimed: Uint128::new(12683916792),
             lockup_end,
-            release_unit: 1,
+            release_unit: Milliseconds::from_seconds(1),
             release_amount: WithdrawalType::Percentage(percent_release),
-            last_claimed_release_time: lockup_end + 2 * release_unit,
+            last_claimed_release_time: lockup_end.plus_seconds(2 * release_unit.seconds()),
         },
         batches().load(deps.as_ref().storage, 1u64).unwrap()
     );
@@ -626,9 +564,9 @@ fn test_claim_batch_not_nice_numbers_multiple_releases() {
             amount: Uint128::new(vesting_amount),
             amount_claimed: Uint128::from(vesting_amount),
             lockup_end,
-            release_unit: 1,
+            release_unit: Milliseconds::from_seconds(1),
             release_amount: WithdrawalType::Percentage(percent_release),
-            last_claimed_release_time: lockup_end + duration + 2,
+            last_claimed_release_time: lockup_end.plus_seconds(duration + 2),
         },
         batches().load(deps.as_ref().storage, 1u64).unwrap()
     );
@@ -640,7 +578,7 @@ fn test_claim_batch_middle_of_interval() {
     init(deps.as_mut());
     let info = mock_info("owner", &coins(100, "uusd"));
 
-    let release_unit = 10;
+    let release_unit = Milliseconds::from_seconds(10);
 
     // Create batch.
     let msg = ExecuteMsg::CreateBatch {
@@ -663,14 +601,14 @@ fn test_claim_batch_middle_of_interval() {
 
     let mut env = mock_env();
     // Only halfway to first release.
-    env.block.time = env.block.time.plus_seconds(release_unit / 2);
+    env.block.time = env.block.time.plus_seconds(release_unit.seconds() / 2);
 
     let res = execute(deps.as_mut(), env.clone(), info.clone(), msg.clone());
 
     assert_eq!(ContractError::WithdrawalIsEmpty {}, res.unwrap_err());
 
     // First release available and halfway to second -> result is rounding down.
-    env.block.time = env.block.time.plus_seconds(release_unit);
+    env.block.time = env.block.time.plus_seconds(release_unit.seconds());
     let res = execute(deps.as_mut(), env, info, msg).unwrap();
 
     assert_eq!(
@@ -685,16 +623,16 @@ fn test_claim_batch_middle_of_interval() {
             .add_attribute("amount_left", "90"),
         res
     );
-    let lockup_end = mock_env().block.time.seconds();
+    let lockup_end = Milliseconds::from_seconds(mock_env().block.time.seconds());
 
     assert_eq!(
         Batch {
             amount: Uint128::new(100),
             amount_claimed: Uint128::new(10),
             lockup_end,
-            release_unit: 10,
+            release_unit: Milliseconds::from_seconds(10),
             release_amount: WithdrawalType::Amount(Uint128::new(10)),
-            last_claimed_release_time: lockup_end + release_unit,
+            last_claimed_release_time: lockup_end.plus_milliseconds(release_unit),
         },
         batches().load(deps.as_ref().storage, 1u64).unwrap()
     );
@@ -706,7 +644,7 @@ fn test_claim_batch_multiple_claims() {
     init(deps.as_mut());
     let info = mock_info("owner", &coins(100, "uusd"));
 
-    let release_unit = 10;
+    let release_unit = Milliseconds::from_seconds(10);
 
     // Create batch.
     let msg = ExecuteMsg::CreateBatch {
@@ -724,7 +662,7 @@ fn test_claim_batch_multiple_claims() {
     let mut env = mock_env();
 
     // 4 releases are available.
-    env.block.time = env.block.time.plus_seconds(4 * release_unit);
+    env.block.time = env.block.time.plus_seconds(4 * release_unit.seconds());
 
     // Claim only the first release.
     let msg = ExecuteMsg::Claim {
@@ -745,7 +683,7 @@ fn test_claim_batch_multiple_claims() {
             .add_attribute("amount_left", "90"),
         res
     );
-    let lockup_end = mock_env().block.time.seconds();
+    let lockup_end = Milliseconds::from_seconds(mock_env().block.time.seconds());
 
     assert_eq!(
         Batch {
@@ -754,7 +692,7 @@ fn test_claim_batch_multiple_claims() {
             lockup_end,
             release_unit,
             release_amount: WithdrawalType::Amount(Uint128::new(10)),
-            last_claimed_release_time: lockup_end + release_unit,
+            last_claimed_release_time: lockup_end.plus_milliseconds(release_unit),
         },
         batches().load(deps.as_ref().storage, 1u64).unwrap()
     );
@@ -778,7 +716,7 @@ fn test_claim_batch_multiple_claims() {
             .add_attribute("amount_left", "60"),
         res
     );
-    let lockup_end = mock_env().block.time.seconds();
+    let lockup_end = Milliseconds::from_seconds(mock_env().block.time.seconds());
 
     assert_eq!(
         Batch {
@@ -787,7 +725,7 @@ fn test_claim_batch_multiple_claims() {
             lockup_end,
             release_unit,
             release_amount: WithdrawalType::Amount(Uint128::new(10)),
-            last_claimed_release_time: lockup_end + 4 * release_unit,
+            last_claimed_release_time: lockup_end.plus_seconds(4 * release_unit.seconds()),
         },
         batches().load(deps.as_ref().storage, 1u64).unwrap()
     );
@@ -799,7 +737,7 @@ fn test_claim_batch_all_releases() {
     init(deps.as_mut());
     let info = mock_info("owner", &coins(100, "uusd"));
 
-    let release_unit = 10;
+    let release_unit = Milliseconds::from_seconds(10);
 
     // Create batch.
     let msg = ExecuteMsg::CreateBatch {
@@ -818,7 +756,7 @@ fn test_claim_batch_all_releases() {
 
     // All releases are available and then some (10 * release_unit would be when all releases
     // become available).
-    env.block.time = env.block.time.plus_seconds(15 * release_unit);
+    env.block.time = env.block.time.plus_seconds(15 * release_unit.seconds());
 
     // Claim only the first release.
     let msg = ExecuteMsg::Claim {
@@ -839,7 +777,7 @@ fn test_claim_batch_all_releases() {
             .add_attribute("amount_left", "0"),
         res
     );
-    let lockup_end = mock_env().block.time.seconds();
+    let lockup_end = Milliseconds::from_seconds(mock_env().block.time.seconds());
 
     assert_eq!(
         Batch {
@@ -848,7 +786,7 @@ fn test_claim_batch_all_releases() {
             lockup_end,
             release_unit,
             release_amount: WithdrawalType::Amount(Uint128::new(10)),
-            last_claimed_release_time: lockup_end + 15 * release_unit,
+            last_claimed_release_time: lockup_end.plus_seconds(15 * release_unit.seconds()),
         },
         batches().load(deps.as_ref().storage, 1u64).unwrap()
     );
@@ -865,7 +803,7 @@ fn test_claim_batch_too_high_of_claim() {
     init(deps.as_mut());
     let info = mock_info("owner", &coins(100, "uusd"));
 
-    let release_unit = 10;
+    let release_unit = Milliseconds::from_seconds(10);
 
     // Create batch.
     let msg = ExecuteMsg::CreateBatch {
@@ -882,7 +820,7 @@ fn test_claim_batch_too_high_of_claim() {
 
     let mut env = mock_env();
     // A single release is available.
-    env.block.time = env.block.time.plus_seconds(release_unit);
+    env.block.time = env.block.time.plus_seconds(release_unit.seconds());
 
     // Try to claim 3 releases.
     let msg = ExecuteMsg::Claim {
@@ -905,16 +843,16 @@ fn test_claim_batch_too_high_of_claim() {
             .add_attribute("amount_left", "90"),
         res
     );
-    let lockup_end = mock_env().block.time.seconds();
+    let lockup_end = Milliseconds::from_seconds(mock_env().block.time.seconds());
 
     assert_eq!(
         Batch {
             amount: Uint128::new(100),
             amount_claimed: Uint128::new(10),
             lockup_end,
-            release_unit: 10,
+            release_unit: Milliseconds::from_seconds(10),
             release_amount: WithdrawalType::Amount(Uint128::new(10)),
-            last_claimed_release_time: lockup_end + release_unit,
+            last_claimed_release_time: lockup_end.plus_milliseconds(release_unit),
         },
         batches().load(deps.as_ref().storage, 1u64).unwrap()
     );
@@ -942,7 +880,7 @@ fn test_claim_all() {
     let mut deps = mock_dependencies_custom(&[]);
     init(deps.as_mut());
 
-    let release_unit = 10;
+    let release_unit = Milliseconds::from_seconds(10);
 
     let release_amount = WithdrawalType::Amount(Uint128::new(10));
     // Create batch.
@@ -952,17 +890,22 @@ fn test_claim_all() {
     create_batch(
         deps.as_mut(),
         None,
-        release_unit / 2,
+        Milliseconds::from_seconds(release_unit.seconds() / 2),
         release_amount.clone(),
     );
 
     // Create batch with a different release_unit scale (not a factor).
-    create_batch(deps.as_mut(), None, 12, release_amount.clone());
+    create_batch(
+        deps.as_mut(),
+        None,
+        Milliseconds::from_seconds(12),
+        release_amount.clone(),
+    );
 
     // Create batch that is still locked up.
     create_batch(
         deps.as_mut(),
-        Some(100),
+        Some(Milliseconds::from_seconds(100)),
         release_unit,
         release_amount.clone(),
     );
@@ -973,7 +916,7 @@ fn test_claim_all() {
 
     // Speed up time.
     let mut env = mock_env();
-    env.block.time = env.block.time.plus_seconds(release_unit * 2);
+    env.block.time = env.block.time.plus_seconds(release_unit.seconds() * 2);
 
     // Query batches
     let msg = QueryMsg::Batches {
@@ -983,7 +926,7 @@ fn test_claim_all() {
     let res: Vec<BatchResponse> =
         from_json(query(deps.as_ref(), env.clone(), msg).unwrap()).unwrap();
 
-    let lockup_end = mock_env().block.time.seconds();
+    let lockup_end = Milliseconds::from_seconds(mock_env().block.time.seconds());
     assert_eq!(
         vec![
             BatchResponse {
@@ -1004,7 +947,7 @@ fn test_claim_all() {
                 amount_available_to_claim: Uint128::new(40),
                 number_of_available_claims: Uint128::new(4),
                 lockup_end,
-                release_unit: release_unit / 2,
+                release_unit: Milliseconds::from_seconds(release_unit.seconds() / 2),
                 release_amount: WithdrawalType::Amount(Uint128::new(10)),
                 last_claimed_release_time: lockup_end,
             },
@@ -1015,7 +958,7 @@ fn test_claim_all() {
                 amount_available_to_claim: Uint128::new(10),
                 number_of_available_claims: Uint128::new(1),
                 lockup_end,
-                release_unit: 12,
+                release_unit: Milliseconds::from_seconds(12),
                 release_amount: WithdrawalType::Amount(Uint128::new(10)),
                 last_claimed_release_time: lockup_end,
             },
@@ -1025,10 +968,10 @@ fn test_claim_all() {
                 amount_claimed: Uint128::zero(),
                 amount_available_to_claim: Uint128::zero(),
                 number_of_available_claims: Uint128::zero(),
-                lockup_end: lockup_end + 100,
+                lockup_end: lockup_end.plus_seconds(100),
                 release_unit,
                 release_amount: WithdrawalType::Amount(Uint128::new(10)),
-                last_claimed_release_time: lockup_end + 100,
+                last_claimed_release_time: lockup_end.plus_seconds(100),
             },
         ],
         res
@@ -1055,7 +998,7 @@ fn test_claim_all() {
         res
     );
 
-    let lockup_end = mock_env().block.time.seconds();
+    let lockup_end = Milliseconds::from_seconds(mock_env().block.time.seconds());
     assert_eq!(
         Batch {
             amount: Uint128::new(100),
@@ -1063,7 +1006,7 @@ fn test_claim_all() {
             lockup_end,
             release_unit,
             release_amount: release_amount.clone(),
-            last_claimed_release_time: lockup_end + release_unit * 2,
+            last_claimed_release_time: lockup_end.plus_seconds(release_unit.seconds() * 2),
         },
         batches().load(deps.as_ref().storage, 1u64).unwrap()
     );
@@ -1073,9 +1016,9 @@ fn test_claim_all() {
             amount: Uint128::new(100),
             amount_claimed: Uint128::new(40),
             lockup_end,
-            release_unit: release_unit / 2,
+            release_unit: Milliseconds::from_seconds(release_unit.seconds() / 2),
             release_amount: release_amount.clone(),
-            last_claimed_release_time: lockup_end + release_unit * 2,
+            last_claimed_release_time: lockup_end.plus_seconds(release_unit.seconds() * 2),
         },
         batches().load(deps.as_ref().storage, 2u64).unwrap()
     );
@@ -1085,9 +1028,9 @@ fn test_claim_all() {
             amount: Uint128::new(100),
             amount_claimed: Uint128::new(10),
             lockup_end,
-            release_unit: 12,
+            release_unit: Milliseconds::from_seconds(12),
             release_amount,
-            last_claimed_release_time: lockup_end + 12,
+            last_claimed_release_time: lockup_end.plus_seconds(12),
         },
         batches().load(deps.as_ref().storage, 3u64).unwrap()
     );