+++ packages/andromeda-fungible-tokens/src/lib.rs
@@ -2,4 +2,5 @@ pub mod airdrop;
 pub mod cw20;
 pub mod cw20_exchange;
 pub mod cw20_staking;
+pub mod ics20;
 pub mod lockdrop;